+++ src/Tokens/AbstractToken.php
@@ -8,13 +8,13 @@ abstract class AbstractToken implements Token
     private $depth;
 
     /** @var int */
-    private $line;
+    protected $line;
 
     /** @var null|Token */
     private $parent;
 
     /** @var int */
-    private $position;
+    protected $position;
 
     /** @var boolean */
     private $throwOnError;
@@ -22,10 +22,13 @@ abstract class AbstractToken implements Token
     /** @var string */
     private $type;
 
+    /** @var string */
+    protected $value;
+
     /**
      * Constructor
      */
-    public function __construct($type, Token $parent = null, $throwOnError = false)
+    public function __construct(string $type, Token $parent = null, bool $throwOnError = false)
     {
         if (!$this->isValidType($type)) {
             throw new \InvalidArgumentException('Invalid type: ' . $type);
@@ -39,11 +42,12 @@ abstract class AbstractToken implements Token
         $this->line = 0;
         $this->position = 0;
         $this->parent = $parent;
-        $this->throwOnError = (boolean) $throwOnError;
+        $this->throwOnError = $throwOnError;
         $this->type = $type;
+        $this->value = '';
     }
 
-    public function getDepth()
+    public function getDepth() : int
     {
         return $this->depth;
     }
@@ -51,22 +55,12 @@ abstract class AbstractToken implements Token
     /**
      * Getter for 'line'.
      */
-    public function getLine()
+    public function getLine() : int
     {
         return $this->line;
     }
 
-    /**
-     * Chainable setter for 'line'.
-     */
-    public function setLine($line)
-    {
-        $this->line = (int) $line;
-
-        return $this;
-    }
-
-    public function isClosingElementImplied($html)
+    public function isClosingElementImplied(string $html) : bool
     {
         return false;
     }
@@ -79,19 +73,9 @@ abstract class AbstractToken implements Token
     /**
      * Getter for 'position'.
      */
-    public function getPosition()
+    public function getPosition() : int
     {
         return $this->position;
-    }
-
-    /**
-     * Chainable setter for 'position'.
-     */
-    public function setPosition($position)
-    {
-        $this->position = (int) $position;
-
-        return $this;
     }
 
     /**
@@ -109,42 +93,52 @@ abstract class AbstractToken implements Token
      *
      * @return string
      */
-    public function getType()
+    public function getType() : string
     {
         return $this->type;
     }
 
-    public function isCDATA()
+    /**
+     * Getter for "value"
+     *
+     * @return string
+     */
+    public function getValue() : string
+    {
+        return $this->value;
+    }
+
+    public function isCDATA() : bool
     {
         return $this->type === Token::CDATA;
     }
 
-    public function isComment()
+    public function isComment() : bool
     {
         return $this->type === Token::COMMENT;
     }
 
-    public function isDocType()
+    public function isDocType() : bool
     {
         return $this->type === Token::DOCTYPE;
     }
 
-    public function isElement()
+    public function isElement() : bool
     {
         return $this->type === Token::ELEMENT;
     }
 
-    public function isPhp()
+    public function isPhp() : bool
     {
         return $this->type === Token::PHP;
     }
 
-    public function isText()
+    public function isText() : bool
     {
         return $this->type === Token::TEXT;
     }
 
-    protected function isValidType($type)
+    protected function isValidType(string $type) : bool
     {
         return $type === Token::CDATA
             || $type === Token::COMMENT